refactor: Change point conversion rate to integer
Updates the data type for the point-to-IDR conversion rate from `Float` to `Int` to simplify calculations and ensure consistency. The conversion rate is now a whole number (1 point = 100 IDR).

- **`Exchange.kt`**: Changed `conversionRate` in the `ExchangeMethod` data class from `Float` to `Int`.
- **`ExchangeRepositoryImpl.kt`**: Updated the hardcoded `conversionRate` for all exchange methods from `100.0f` to `100`.
- **`ExchangePointViewModel.kt`**: Modified the `amountReceived` calculation to use integer arithmetic.
- **`ExchangePointScreen.kt`**: Adjusted the calculation for `amountReceived` in the confirmation dialog to align with the new integer-based logic.

+++ app/src/main/java/com/ahargunyllib/growth/model/Exchange.kt
@@ -20,7 +20,7 @@ data class ExchangeMethod(
     val iconResId: Int = 0,
     val minAmount: Int = 100, // Minimum points to exchange
     val maxAmount: Int = 10000, // Maximum points per transaction
-    val conversionRate: Float = 100.0f, // Points to IDR ratio (100 points = 10,000 IDR = 1 point = 100 IDR)
+    val conversionRate: Int = 100, // Points to IDR ratio (1 point = 100 IDR)
     val adminFee: Int = 0, // Admin fee in points
     val isActive: Boolean = true
 )
